requirements.txt
lmdb==1.3.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
pytorch_lightning==1.6.0
PyYAML==6.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
